Create class `TreePathUtil`
